Arch: import IFC, move methods used by import and export to helper
