+++ src/Mod/Arch/exportIFC.py
@@ -39,10 +39,11 @@ import Draft
 import Arch
 import DraftVecUtils
 import ArchIFCSchema
-from DraftGeomUtils import vec
-from importIFC import dd2dms
-from importIFC import decode
 import exportIFCHelper
+
+from DraftGeomUtils import vec
+from importIFCHelper import dd2dms
+from importIFCHelper import decode
 
 
 ## @package exportIFC
